+++ ContactMaster/Controllers/LoginController.cs
@@ -1,4 +1,5 @@
-﻿using ContactMasterService;
+﻿using ContactMaster.Services;
+using ContactMasterService;
 using Dominio.Interfaces;
 using Dominio.Models;
 using Microsoft.AspNetCore.Mvc;
@@ -10,10 +11,10 @@ namespace ContactMaster.Controllers
     public class LoginController : Controller
     {
         private readonly IUsuarioRepositorio _usuarioRepositorio;
-        private readonly ISessao _sessao;
-        private readonly IEmail _email;
+        private readonly ISessaoService _sessao;
+        private readonly IEmailService _email;
 
-        public LoginController(IUsuarioRepositorio usuarioRepositorio, ISessao sessao, IEmail email)
+        public LoginController(IUsuarioRepositorio usuarioRepositorio, ISessaoService sessao, IEmailService email)
         {
             _usuarioRepositorio = usuarioRepositorio;
             _sessao = sessao;
