+++ ContactMaster/Controllers/ModificarSenhaController.cs
@@ -1,4 +1,5 @@
-﻿using Dominio.Interfaces;
+﻿using ContactMaster.Services;
+using Dominio.Interfaces;
 using Dominio.Models;
 using Microsoft.AspNetCore.Mvc;
 using System;
@@ -9,9 +10,9 @@ namespace ContactMaster.Controllers
     public class ModificarSenhaController : Controller
     {
         private readonly IUsuarioRepositorio _usuarioRepositorio;
-        private readonly ISessao _sessao;
+        private readonly ISessaoService _sessao;
 
-        public ModificarSenhaController(IUsuarioRepositorio usuarioRepositorio, ISessao sessao)
+        public ModificarSenhaController(IUsuarioRepositorio usuarioRepositorio, ISessaoService sessao)
         {
             _usuarioRepositorio = usuarioRepositorio;
             _sessao = sessao;
