+++ ContactMasterService/Interfaces/IBonusService.cs
