+++ ContactMasterService/Interfaces/IUsuarioService.cs
