+++ ContactMasterService/Services/LoginService.cs
@@ -1,15 +1,16 @@
-﻿using Dominio.Interfaces;
+﻿using ContactMaster.Services;
+using Dominio.Interfaces;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace ContactMasterService
+namespace ContactMasterService.Services
 {
     public class LoginService : ILoginService
     {
-        
+
         private readonly ILoginRepositorio _loginRepositorio;
 
         public LoginService(ILoginRepositorio loginRepositorio)
