+++ ContactMasterService/Services/UsuarioService.cs
@@ -7,7 +7,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace ContactMasterService
+namespace ContactMasterService.Services
 {
     public class UsuarioService : IUsuarioService
     {
