--- Dominio/Interfaces/ILoginService.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Dominio.Interfaces
-{
-    public interface ILoginService
-    {
-        Task<bool> Login(string email, string password);
-    }
-}
