+++ MasterApi/Controllers/ContatoController.cs
@@ -1,5 +1,6 @@
 ﻿using ContactMaster.Services;
 using ContactMasterService;
+using ContactMasterService.Services;
 using Dominio.Models;
 using Microsoft.AspNetCore.Mvc;
 using System.Collections.Generic;
