+++ MasterApi/Controllers/HomeController.cs
@@ -1,5 +1,6 @@
 ﻿using ContactMaster.Models;
 using ContactMasterService;
+using ContactMasterService.Services;
 using Microsoft.AspNetCore.Mvc;
 using System.Diagnostics;
